Improving vim markdown support
Just adding the autocmd to the ftplugin worked when we were only editing
markdown files. However if we opened up a non-markdown file in another
buffer it would try to format that file on save.

The new approach is to use an autogrp to enable the functionality only
for markdown files.